dv_macropad/requirements.txt
pyautogui==0.9.54
pillow>=11,<12
pyinstaller>=6.15,<7
